Update WalletContractV1 identifier types
Fixes #9

+++ fakewallet/src/main/java/com/solanamobile/fakewallet/ui/AccountListAdapter.kt
@@ -68,7 +68,7 @@ class AccountListAdapter(
 
     override fun getItemId(position: Int): Long {
         val account = getItem(position)
-        return account.id.toLong()
+        return account.id
     }
 }
